Separating RealmStorage
