+++ flmodules/src/flo/mod.rs
@@ -4,6 +4,7 @@ pub mod flo;
 pub mod realm;
 
 // Easy interaction with dht_storage and realms
+pub mod realm_storage;
 pub mod realm_view;
 
 pub mod storage;
